% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_typology.R
\name{add_typologies}
\alias{add_typologies}
\title{Add freshwater typologies to sampling sites}
\usage{
add_typologies(data)
}
\arguments{
\item{data}{data.table}
}
\value{
data.table
}
\description{
Add freshwater typologies to sampling sites
}
\examples{

}
