% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_strength.R
\name{classification_strength}
\alias{classification_strength}
\title{Compute classification strength of clustering}
\usage{
classification_strength(dist, grouping, season, typology, permutations)
}
\arguments{
\item{dist}{distance matrix}

\item{grouping}{character vector with group membership of sites}

\item{season}{character, for which season is the test run?}

\item{typology}{character, for this typology is this test run?}

\item{permutations}{}
}
\value{
data.table
}
\description{
Computes the classification strength of a classification.
}
