% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_typical_comm.R
\name{compute_typical_comm}
\alias{compute_typical_comm}
\title{Compute typical communities}
\usage{
compute_typical_comm(community, grouping, out = "similarity", threshold = 0.66)
}
\arguments{
\item{community}{data.frame with sites as rows and taxa as columns}

\item{grouping}{character vector with group membership of sites}

\item{out}{character. Prefered output. Either the mean similarity between typical communities or list of typical communities.}

\item{threshold}{numeric. Relative frequency a taxon must attain to be regarded as common.}
}
\value{
Either the mean similarity between typical communities or list of typical communities.
}
\description{
Compute typical communities
}
