% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_prepare_data_2.R
\name{ct_prepare_data2}
\alias{ct_prepare_data2}
\title{Compute distance matrix from result of ct_preparedata1}
\usage{
ct_prepare_data2(data, ncores = 1, distance_metric = "binary")
}
\arguments{
\item{data}{data.table}

\item{ncores}{integer. Number of cores to be used in the computation of the distance matrix}

\item{distance_metric}{character. What distance metric should be used in the distance matrix}
}
\value{
dist.
}
\description{
Compute distance matrix from result of ct_preparedata1
}
