% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_taxon_table.R
\name{fill_taxon_table}
\alias{fill_taxon_table}
\title{Fill taxon table}
\usage{
fill_taxon_table(
  o,
  s = NULL,
  g = NULL,
  f = NULL,
  or = NULL,
  sc = NULL,
  c = NULL,
  p = NULL,
  k = NULL
)
}
\arguments{
\item{o}{character. New order name.}

\item{s}{character. New species name.}

\item{g}{character. New genus name.}

\item{f}{character. New family name.}

\item{sc}{character. New subclass name.}

\item{c}{character. New class name.}

\item{p}{character. New phylum name.}

\item{k}{character. New kingdom name.}
}
\value{
data.table
}
\description{
Fill taxon table
}
