% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indvalstat.R
\name{indvalstat}
\alias{indvalstat}
\title{Compute indicator value statistics}
\usage{
indvalstat(community, grouping, perm, typology, season)
}
\arguments{
\item{community}{data.frame with sites as rows and taxa as columns}

\item{grouping}{character vector with group membership of sites}

\item{perm}{number of permutations to compute p-value of number of indicators and mean indicator statistic. NOT permutations in the call to multipatt.}

\item{typology}{character, for this typology is this test run?}

\item{season}{character, for which season is the test run?}
}
\value{
data.table with typology, season, variable (number of indicator taxa or mean indicator value of significant indicator taxa), test statistic values and permutation p_vaules.
}
\description{
This function returns the number of indicator taxa and the mean indicator value of significant indicator taxa.
To compute indicator values the indicspec::multipatt() with 999 permutations and the group-equalized static are used.
}
