% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_most_recent_date.R
\name{most_recent_date}
\alias{most_recent_date}
\title{Get most recent date for files with same name}
\usage{
most_recent_date(folder, file)
}
\arguments{
\item{folder}{character; path to the file.}

\item{file}{character; name of file}
}
\value{
character; most recent date
}
\description{
If you have multiple files with the same name but different dates as part of the name, e.g. 21-05-03_test.rds and 21-05-04_test.rds this function will extract the most recent date.
}
