% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newst_samples.R
\name{newest_sample}
\alias{newest_sample}
\title{Reduce a data set with multiple samples per site to the most recent sample for each site.}
\usage{
newest_sample(x)
}
\arguments{
\item{x}{data set to reduce}
}
\value{
x data set with only the most recent samples for each site
}
\description{
Reduce a data set with multiple samples per site to the most recent sample for each site.
}
