% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{nop}
\alias{nop}
\title{Function Returning Nothing}
\usage{
nop(x = NULL)
}
\arguments{
\item{x}{R object to be returned invisibly.}
}
\description{
Return nothing via \code{\link[base]{invisible}()}.
}
\examples{
nop()

}
\seealso{
\code{\link[base]{invisible}}
}
