% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typical_comm.R
\name{typical_comm}
\alias{typical_comm}
\title{Compute statistical significance of typical communities}
\usage{
typical_comm(community, grouping, perm, typology, season, threshold = 0.5)
}
\arguments{
\item{community}{data.frame with sites as rows and taxa as columns}

\item{grouping}{character vector with group membership of sites}

\item{perm}{number of permutations to compute p-value of number of indicators and mean indicator statistic. NOT permutations in the call to multipatt.}

\item{typology}{character, for this typology is this test run?}

\item{season}{character, for which season is the test run?}

\item{threshold}{numeric. Relative frequency a taxon must attain to be regarded as common.}
}
\value{
data.table with mean similarity and p-value.
}
\description{
Compute statistical significance of typical communities
}
