% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_taxonomy.R
\name{update_taxonomy}
\alias{update_taxonomy}
\title{Update the taxonomy table}
\usage{
update_taxonomy(TU)
}
\arguments{
\item{TU}{character vector}
}
\value{
data.frame; updated taxontable.
}
\description{
Update the taxontable with the taxa in TU. The taxontable must already be loaded.
}
