% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_taxonomy2.R
\name{update_taxonomy2}
\alias{update_taxonomy2}
\title{Update the taxonomy table2}
\usage{
update_taxonomy2(TU)
}
\arguments{
\item{TU}{character vector}
}
\value{
data.frame; updated taxontable.
}
\description{
Update the taxontable with the taxa in TU. The taxontable must already be loaded.
Altered version of the first function is was keept intact to keep old scirpts running.
}
