% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_prepare_data_1.R
\name{ct_prepare_data1}
\alias{ct_prepare_data1}
\title{Prepare data for comparison of typologies}
\usage{
ct_prepare_data1(data, taxon, abundance.to.pa = TRUE, typologies)
}
\arguments{
\item{data}{data.table}

\item{taxon}{character}

\item{abundance.to.pa}{boolean}
}
\value{
data.table
}
\description{
Prepare data for comparison of typologies
}
\examples{

}
