% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_taxon_table.R
\name{fill_taxon_table}
\alias{fill_taxon_table}
\title{Fill taxon table}
\usage{
fill_taxon_table(o, a, b, c, d, e, f, g)
}
\arguments{
\item{o}{character. Original Name.}

\item{a}{character. New species name.}

\item{b}{character. New genus name.}

\item{c}{character. New family name.}

\item{d}{character. New order name.}

\item{e}{character. New subclass name.}

\item{f}{character. New class name.}

\item{g}{character. New phylum name.}
}
\value{
data.table
}
\description{
Fill taxon table
}
